Move elftypes to lib/
